+++ editor/plugins/input_event_editor_plugin.cpp
@@ -30,6 +30,9 @@
 
 #include "input_event_editor_plugin.h"
 
+#include "editor/event_listener_line_edit.h"
+#include "editor/input_event_configuration_dialog.h"
+
 void InputEventConfigContainer::_bind_methods() {
 }
